Separate `Server` from `Connection`
The only parts of the interface they share are the `close` operation and the
`sock` and `addr` fields. They have very little in common and I think the
inheritance is kind of forced just for a minor convenience and not serving
a useful purpose. To make sure this doesn't develop into a larger inheritance
mess, I'd like to separate these now. This also puts `Server` in its own files
for better organization and faster compilation.
